restructured into packages, fixed dependencies
