+++ item/item_resource.go
@@ -1,0 +1,1 @@
+package item
