+++ user/user.go
@@ -1,4 +1,4 @@
-package main
+package user
 
 import (
 	"gopkg.in/mgo.v2/bson"
