+++ user/user_storage.go
@@ -1,6 +1,8 @@
-package main
+package user
 
 import (
+	"github.com/beisner/OBO/util"
+
 	"gopkg.in/mgo.v2"
 	"gopkg.in/mgo.v2/bson"
 )
@@ -11,7 +13,7 @@ type UserStorage struct {
 	col *mgo.Collection // collection right from mongo
 }
 
-func NewUserStorage(db *mgo.Database) *UserStorage {
+func newUserStorage(db *mgo.Database) *UserStorage {
 	us := new(UserStorage)
 	us.db = db
 	us.col = db.C("user")
@@ -20,36 +22,36 @@ func NewUserStorage(db *mgo.Database) *UserStorage {
 
 func (u *UserStorage) ExistsUser(id bson.ObjectId) bool {
 	n, err := u.col.FindId(id).Count()
-	logerr(err)
+	util.Logerr(err)
 	return n > 0
 }
 
 func (u *UserStorage) GetUser(id bson.ObjectId) *User {
 	result := User{}
-	logerr(u.col.FindId(id).One(&result))
+	util.Logerr(u.col.FindId(id).One(&result))
 	return &result
 }
 
 func (u *UserStorage) InsertUser(user User) (bool, bson.ObjectId) {
 	user.Id = bson.NewObjectId()
-	logerr(u.col.Insert(user))
+	util.Logerr(u.col.Insert(user))
 	return true, user.Id
 }
 
 func (u *UserStorage) UpdateUser(user User) bool {
-	logerr(u.col.UpdateId(user.Id, user))
+	util.Logerr(u.col.UpdateId(user.Id, user))
 	return true
 }
 
 func (u *UserStorage) DeleteUser(id bson.ObjectId) *User {
 	result := User{}
-	logerr(u.col.FindId(id).One(&result))
-	logerr(u.col.RemoveId(id))
+	util.Logerr(u.col.FindId(id).One(&result))
+	util.Logerr(u.col.RemoveId(id))
 	return &result
 }
 
 func (u *UserStorage) Length() int {
 	n, err := u.col.Count()
-	logerr(err)
+	util.Logerr(err)
 	return n
 }
